FEAT: Add QC metrics to both pipeline results
Pharmcat and lookup pipelines need to include the QC metrics QUAL,
FILTER, FORMAT/DP, FORMAT/GQ, FORMAT/MQ and FORMAT/QD with their regular
results, for use in adding warnings to the frontend.

Add the additional QC metrics, using "." if the format fields are not
defined in the VCF header.