FIX: Update lookup to work with test case
The search for format tags doesn't use the LOCAL_DIR environment
variable, so test case VCFs are not found. In addition, the fields added
by a442941 have not also been added to the test case.

Use LOCA_DIR as the CWD for seaching the vcf header, so the test
directory works as expected. Also add the values for the additional
fields present in the test VCF, and "." for those not present.